requirements.txt
amqp==5.0.7
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.1
decorator==5.1.0
Deprecated==1.2.13
Django==3.2.10
django-rest-knox==4.1.0
djangorestframework==3.13.1
dynaconf==3.1.7
flake8==4.0.1
greenlet==1.1.2
hvac==0.11.2
idna==3.3
Jinja2==3.0.3
knox==0.1.14
kombu==5.2.2
loguru==0.5.3
MarkupSafe==2.0.1
mccabe==0.6.1
plaid-python==5.0.0
prompt-toolkit==3.0.24
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.26.0
shortuuid==1.0.8
six==1.16.0
SQLAlchemy==1.4.28
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
validators==0.18.2
vine==5.0.0
wcwidth==0.2.5
win32-setctime==1.0.4
wrapt==1.13.3
